// wpref-frontend/src/app/api/generated/model/token-obtain-pair.ts
/**
 * WpRef API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenObtainPairDto { 
    username: string;
    password: string;
    readonly access: string;
    readonly refresh: string;
}