// wpref-frontend/src/app/api/generated/model/custom-user-create.ts
/**
 * WpRef API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageEnumDto } from './language-enum';


export interface CustomUserCreateDto { 
    /**
     * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
     */
    username: string;
    email?: string;
    first_name?: string;
    last_name?: string;
    password: string;
    language?: LanguageEnumDto;
}